+++ dev/e2e_app/lib/keys.dart
@@ -49,6 +49,7 @@ class Keys {
   static const microphonePermissionTile = Key('microphone');
   static const locationPermissionTile = Key('location');
   static const galleryPermissionTile = Key('gallery');
+  static const batteryPermissionTile = Key('battery');
   static const permissionsScreen = Key('permissionsScreen');
   static const requestCameraPermissionButton = Key(
     'requestCameraPermissionButton',
@@ -61,6 +62,9 @@ class Keys {
   );
   static const requestGalleryPermissionButton = Key(
     'requestGalleryPermissionButton',
+  );
+  static const requestBatteryPermissionButton = Key(
+    'requestBatteryPermissionButton',
   );
 
   // location screen